Resolved merge conflicts before merging into develop

+++ src/main/java/ons/group8/services/AdminService.java
@@ -13,6 +13,6 @@ public interface AdminService {
 
     Optional<User> findUsersByEmail(String email);
 
-    Optional<Role> findRolesById(int id);
+    Optional<Role> findRolesById(Long id);
 
 }
